+++ lib/src/typing_text.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 
 class TypingText extends StatefulWidget {
@@ -12,22 +11,18 @@ class TypingText extends StatefulWidget {
   final Duration minDelay;
   final Duration maxDelay;
   final VoidCallback? onFinished;
-  final bool simulateTypos;
-  final bool enableSound;
   final bool useMarkdown;
 
   const TypingText({
+    Key? key,
     this.textStream,
     this.text,
     this.style,
     this.minDelay = const Duration(milliseconds: 30),
     this.maxDelay = const Duration(milliseconds: 90),
     this.onFinished,
-    this.simulateTypos = true,
-    this.enableSound = false,
     this.useMarkdown = false,
-    super.key,
-  });
+  }) : super(key: key);
 
   @override
   State<TypingText> createState() => _TypingTextState();
@@ -36,10 +31,10 @@ class TypingText extends StatefulWidget {
 class _TypingTextState extends State<TypingText> {
   final Random _random = Random();
   String _visibleText = "";
-
   int _charIndex = 0;
   String _targetText = "";
   StreamSubscription<String>? _streamSub;
+  bool _isTyping = false;
 
   @override
   void initState() {
@@ -47,65 +42,47 @@ class _TypingTextState extends State<TypingText> {
 
     if (widget.textStream != null) {
       _streamSub = widget.textStream!.listen((newText) {
-        setState(() {
-          _targetText = newText;
-          _charIndex = 0;
-          _visibleText = "";
-        });
-        _startTyping();
+        _resetAndStart(newText);
       });
     } else if (widget.text != null) {
-      _targetText = widget.text!;
-      _startTyping();
+      _resetAndStart(widget.text!);
     }
   }
 
+  void _resetAndStart(String text) {
+    setState(() {
+      _targetText = text;
+      _charIndex = 0;
+      _visibleText = "";
+    });
+    if (!_isTyping) _startTyping();
+  }
+
   Future<void> _startTyping() async {
-    while (_charIndex < _targetText.length) {
+    _isTyping = true;
+
+    while (_charIndex < _targetText.length && mounted) {
       await Future.delayed(_randomDuration());
 
-      if (!mounted) return;
-
       setState(() {
-        String nextChar = _targetText[_charIndex];
-
-        if (widget.simulateTypos && _random.nextDouble() < 0.05) {
-          // Simulate a typo: insert wrong char
-          String typoChar = String.fromCharCode(_random.nextInt(26) + 97);
-          _visibleText += typoChar;
-          _playSound();
-        } else {
-          _visibleText += nextChar;
-          _charIndex++;
-          _playSound();
-        }
+        _visibleText += _targetText[_charIndex];
+        _charIndex++;
       });
-
-      if (widget.simulateTypos &&
-          _visibleText.length > 1 &&
-          _random.nextDouble() < 0.05) {
-        await Future.delayed(const Duration(milliseconds: 100));
-        setState(() {
-          _visibleText = _visibleText.substring(0, _visibleText.length - 1);
-        });
-        await Future.delayed(_randomDuration());
-      }
     }
 
+    _isTyping = false;
     widget.onFinished?.call();
   }
 
   Duration _randomDuration() {
-    return Duration(
-      milliseconds:
-          widget.minDelay.inMilliseconds +
-          _random.nextInt(widget.maxDelay.inMilliseconds + 1),
-    );
-  }
+    final min = widget.minDelay.inMilliseconds;
+    final max = widget.maxDelay.inMilliseconds;
 
-  void _playSound() {
-    if (!widget.enableSound) return;
-    SystemSound.play(SystemSoundType.alert);
+    if (max <= min) {
+      return Duration(milliseconds: min);
+    }
+
+    return Duration(milliseconds: min + _random.nextInt((max - min) + 1));
   }
 
   @override
@@ -118,9 +95,15 @@ class _TypingTextState extends State<TypingText> {
   Widget build(BuildContext context) {
     return widget.useMarkdown
         ? MarkdownBody(
-          data: _visibleText,
-          styleSheet: MarkdownStyleSheet(p: widget.style ?? const TextStyle()),
-        )
-        : Text(_visibleText, style: widget.style);
+            data: _visibleText,
+            styleSheet:
+                MarkdownStyleSheet.fromTheme(Theme.of(context)).copyWith(
+              p: widget.style ?? const TextStyle(),
+            ),
+          )
+        : Text(
+            _visibleText,
+            style: widget.style,
+          );
   }
 }